Drunk coding in wrong branch + never cared less

+++ Web/Requests/EmployeeRequests/UpdateEmployeeApiRequest.cs
@@ -5,9 +5,9 @@ public class UpdateEmployeeApiRequest
     public required string Name { get; set; }
     public required string Phone { get; set; }
     public bool IsDriver { get; set; }
-    public uint? Passport { get; set; }
+    public ulong? Passport { get; set; }
     public DateTimeOffset DateOfBirth { get; set; }
-    public uint? INN { get; set; }
+    public ulong? INN { get; set; }
     public ulong? AccountNumber { get; set; }
-    public uint? BIK { get; set; }
+    public ulong? BIK { get; set; }
 }
