+++ Web/Responses/EmployeeResponses/ApiEmployee.cs
@@ -6,9 +6,9 @@ public class ApiEmployee
     public required string Name { get; set; }
     public required string Phone { get; set; }
     public bool IsDriver { get; set; }
-    public uint? Passport { get; set; }
+    public ulong? Passport { get; set; }
     public DateTimeOffset DateOfBirth { get; set; }
-    public uint? INN { get; set; }
+    public ulong? INN { get; set; }
     public ulong? AccountNumber { get; set; }
-    public uint? BIK { get; set; }
+    public ulong? BIK { get; set; }
 }
